client/requirements.txt
a2a-sdk
httpx
asyncio
